refactor: useColorThief으로 분리
