+++ src/hooks/index.ts
@@ -10,3 +10,4 @@ export * from './useBookList';
 export * from './useRandomTitle';
 export * from './useGrid';
 export * from './useDebounce';
+export * from './useColorThief';
